test: Move incremental eval compilation tests to subscription module
